fixes issue display of picture from LAC
rewrites the utility method that writes the http stream to a temp file.
Thsi is because the previous routine was not completing the write process
and would leave the file incomplete leading to json parsing errors in the
calling routines.

the change to the Synchronization manager is to delete the temporarly
cache file before the fresh download of keywords is written to the temp
cache.