adds feature to tag content as favourite
This fix introduce a feature in the SearchItem view activity with a rating
icon that when the user clicks tags the content as favourite or if it was
already tagged, it untagged.

When the search item is tagged as favourite, a favourite record is created
in the favourite_record sqlite table.